Refactor models and namespaces for clarity
- Updated namespace from `V275_REST_lib.Models` to `V5_REST_Lib.Models` across various classes, aligning with the new version of the REST library.
- Transitioned from `JObject` to specific model types (`Config` and `ResultsAlt`) in `ImageResult`, `V5Result`, and `ImageResultEntry` classes for stronger typing.
- Modified method signatures in `V5CreateSectorsImageOverlay` and various classes to accept `ResultsAlt` instead of `JObject`, reflecting the move towards strongly typed models.
- Adjusted data binding in `SectorDifferences.xaml` and updated property types in `Grade` and `ValueResult` classes to `double` for increased precision.
- Changed the data model in `Report` and `Sector` classes to use `ResultsAlt.Decodedata`, indicating a shift to a more detailed or differently structured report data model.
- Overall, these changes aim to improve code clarity, type safety, and ease of maintenance by using more specific, strongly typed models for REST API responses.

+++ LabelVal/Sectors/Interfaces/ISectorDifferences.cs
@@ -45,11 +45,11 @@ public class GradeValue
 public class Grade
 {
     public string Name { get; set; }
-    public float Value { get; set; }
+    public double Value { get; set; }
     public string Letter { get; set; }
 
 
-    public Grade(string name, float value, string letter)
+    public Grade(string name, double value, string letter)
     {
         Value = value;
         Letter = letter;
@@ -68,10 +68,10 @@ public class Grade
 public class ValueResult
 {
     public string Name { get; set; }
-    public float Value { get; set; }
+    public double Value { get; set; }
     public string Result { get; set; }
 
-    public ValueResult(string name, float value, string result)
+    public ValueResult(string name, double value, string result)
     {
         Value = value;
         Result = result;
